+++ app/common/types/cennznet-node.types.js
@@ -10,8 +10,8 @@ export const PreDefinedAssetId = {
 export const CustomTokenAssetId = ['100'];
 
 export const PreDefinedAssetIdName = {
-  '0': 'Stake Token',
-  '10': 'Spending Token',
+  '0': 'CENNZ(Stake Token)',
+  '10': 'CPAY(Spending Token)',
   '100': 'Sylo Token',
 };
 
@@ -24,6 +24,11 @@ export const NetworkNameMapping = {
 export const CENNZScanTxUrl = {
   'rimu': 'https://service.centrality.me/cennzscan/tx',
   'kauri': 'https://service.centrality.me/cennzscan/tx', // TODO CENNZScan not support DEV yet
+};
+
+export const CENNZScanAddressUrl = {
+  'rimu': 'https://service.centrality.me/cennzscan/addresses',
+  'kauri': 'https://service.centrality.me/cennzscan/addresses', // TODO CENNZScan not support DEV yet
 };
 
 export const PreDefinedAssetIdObj = {