+++ CHANGELOG.md
@@ -3,10 +3,15 @@
 # vNext
 
 ### Features
+- add CENNZScan address link to account public address
 
 ### Fixes
 
 ### Improvements
+- tidy up asset table layout
+- update background gradient settings
+- update field text bg color to match with design
+- update sync info layout
 
 # v0.5.0 (2019.02.20)